moved functions into relevant test files instead of in utils
