+++ src/test/components/elements/Dashboard.test.tsx
@@ -1,11 +1,60 @@
 import { mockNewsPods, mockPromoPods } from "../../../mocks/data";
 import { Dashboard } from "../../../app/components/elements/Dashboard";
-import { TestUserRole, checkDashboardButtons, renderTestEnvironment } from "../../utils";
-import { screen, waitFor } from "@testing-library/react";
+import { TestUserRole, renderTestEnvironment } from "../../utils";
+import { screen, waitFor, within } from "@testing-library/react";
 import { USER_ROLES, UserRole } from "../../../IsaacApiTypes";
 
 const mockPromoItem = mockPromoPods.results[0];
 const mockFeaturedNewsItem = mockNewsPods.results[1];
+
+const findDashboardButtons = () => {
+  const allUserButtons = screen.getByTestId("show-me-buttons");
+  const dashboardButtons = within(allUserButtons).getAllByRole("link");
+  const teacherButtons = screen.queryByTestId("teacher-dashboard-buttons");
+  const teacherDashboardButtons = teacherButtons ? within(teacherButtons).getAllByRole("link") : null;
+  return { dashboardButtons, teacherDashboardButtons };
+};
+
+export const checkDashboardButtons = (role?: "TEACHER") => {
+  const { dashboardButtons, teacherDashboardButtons } = findDashboardButtons();
+
+  const expectCommonButtons = () => {
+    const commonButtonDetails = [
+      { text: "GCSE resources", href: "/topics/gcse" },
+      { text: "A Level resources", href: "/topics/a_level" },
+      { text: "Events", href: "/events" },
+    ];
+
+    commonButtonDetails.forEach(({ text, href }, index) => {
+      expect(dashboardButtons[index]).toHaveTextContent(text);
+      expect(dashboardButtons[index]).toHaveAttribute("href", href);
+    });
+  };
+
+  const expectTeacherButtons = () => {
+    expect(teacherDashboardButtons!.length).toBe(3);
+    const teacherButtonDetails = [
+      { text: "Key stage 3 courses", href: "https://teachcomputing.org/courses?level=Key+stage+3" },
+      { text: "Key stage 4 courses", href: "https://teachcomputing.org/courses?level=Key+stage+4" },
+      { text: "A level courses", href: "https://teachcomputing.org/courses?level=Post+16" },
+    ];
+
+    teacherButtonDetails.forEach(({ text, href }, index) => {
+      expect(teacherDashboardButtons![index]).toHaveTextContent(text);
+      expect(teacherDashboardButtons![index]).toHaveAttribute("href", href);
+      const buttonIcon = within(teacherDashboardButtons![index]).getByRole("img");
+      expect(buttonIcon).toBeVisible();
+    });
+  };
+
+  expectCommonButtons();
+
+  if (role === "TEACHER") {
+    expectTeacherButtons();
+  } else {
+    expect(teacherDashboardButtons).toBeNull();
+  }
+};
 
 describe("Dashboard", () => {
   const setupTest = (role: TestUserRole, props = {}) => {
