+++ src/test/utils.tsx
@@ -161,13 +161,6 @@ export const checkPageTitle = (pageTitle: string) => {
   expect(title).toBeInTheDocument();
 };
 
-export const checkPasswordInputTypes = (expectedType: string) => {
-  const formFields = getFormFields();
-  const passwordInput = formFields.password() as HTMLInputElement;
-  const confirmPasswordInput = formFields.confirmPassword() as HTMLInputElement;
-  [passwordInput, confirmPasswordInput].forEach((input) => expect(input.type).toBe(expectedType));
-};
-
 export const clickButton = async (buttonName: string) => {
   const button = screen.getByRole("button", { name: buttonName });
   await userEvent.click(button);
@@ -281,119 +274,5 @@ export const fillFormCorrectly = async (correctly: boolean, role: "teacher" | "s
       await userEvent.click(recaptcha());
       break;
     }
-  }
-};
-
-export const extraDownloadEndpoints = {
-  working: [
-    rest.get(API_PATH + `/documents/content/books/gcse_book_23/isaac_cs_gcse_book_2023.pdf`, (req, res, ctx) => {
-      return res(ctx.status(200), ctx.json({ data: "this is a book" }));
-    }),
-    rest.get(API_PATH + `/documents/content/books/workbook_20_aqa/isaac_cs_aqa_book_2022.pdf`, (req, res, ctx) => {
-      return res(ctx.status(200), ctx.json({ data: "this is a book" }));
-    }),
-    rest.get(API_PATH + `/documents/content/books/workbook_20_ocr/isaac_cs_ocr_book_2022.pdf`, (req, res, ctx) => {
-      return res(ctx.status(200), ctx.json({ data: "this is a book" }));
-    }),
-  ],
-  broken: [
-    rest.get(API_PATH + `/documents/content/books/gcse_book_23/isaac_cs_gcse_book_2023.pdf`, (req, res, ctx) => {
-      return res(
-        ctx.status(404),
-        ctx.json({
-          responseCode: 404,
-          responseCodeType: "Not Found",
-          errorMessage: "Unable to locate the file: content/books/gcse_book_23/isaac_cs_gcse_book_2023.pdf.",
-          bypassGenericSiteErrorPage: false,
-        }),
-      );
-    }),
-    rest.get(API_PATH + `/documents/content/books/workbook_20_ocr/isaac_cs_ocr_book_2022.pdf`, (req, res, ctx) => {
-      return res(
-        ctx.status(404),
-        ctx.json({
-          responseCode: 404,
-          responseCodeType: "Not Found",
-          errorMessage: "Unable to locate the file: content/books/workbook_20_ocr/isaac_cs_ocr_book_2022.pdf.",
-          bypassGenericSiteErrorPage: false,
-        }),
-      );
-    }),
-    rest.get(API_PATH + `/documents/content/books/workbook_20_ocr/isaac_cs_ocr_book_2022`, (req, res, ctx) => {
-      return res(
-        ctx.status(404),
-        ctx.json({
-          responseCode: 404,
-          responseCodeType: "Not Found",
-          errorMessage: "Unable to locate the file: /content/books/workbook_20_ocr/isaac_cs_ocr_book_2022.",
-          bypassGenericSiteErrorPage: false,
-        }),
-      );
-    }),
-  ],
-};
-
-export const getDownloadButtons = () => {
-  const gcseButton = () =>
-    screen.getByRole("button", {
-      name: /download gcse workbook/i,
-    });
-  const aqaButton = () =>
-    screen.getByRole("button", {
-      name: /download aqa workbook/i,
-    });
-  const ocrButton = () =>
-    screen.getByRole("button", {
-      name: /download ocr workbook/i,
-    });
-  return { gcseButton, aqaButton, ocrButton };
-};
-
-const findDashboardButtons = () => {
-  const allUserButtons = screen.getByTestId("show-me-buttons");
-  const dashboardButtons = within(allUserButtons).getAllByRole("link");
-  const teacherButtons = screen.queryByTestId("teacher-dashboard-buttons");
-  const teacherDashboardButtons = teacherButtons ? within(teacherButtons).getAllByRole("link") : null;
-  return { dashboardButtons, teacherDashboardButtons };
-};
-
-export const checkDashboardButtons = (role?: "TEACHER") => {
-  const { dashboardButtons, teacherDashboardButtons } = findDashboardButtons();
-
-  const expectCommonButtons = () => {
-    const commonButtonDetails = [
-      { text: "GCSE resources", href: "/topics/gcse" },
-      { text: "A Level resources", href: "/topics/a_level" },
-      { text: "Events", href: "/events" },
-    ];
-
-    commonButtonDetails.forEach(({ text, href }, index) => {
-      expect(dashboardButtons[index]).toHaveTextContent(text);
-      expect(dashboardButtons[index]).toHaveAttribute("href", href);
-    });
-  };
-
-  const expectTeacherButtons = () => {
-    expect(teacherDashboardButtons!.length).toBe(3);
-    const teacherButtonDetails = [
-      { text: "Key stage 3 courses", href: "https://teachcomputing.org/courses?level=Key+stage+3" },
-      { text: "Key stage 4 courses", href: "https://teachcomputing.org/courses?level=Key+stage+4" },
-      { text: "A level courses", href: "https://teachcomputing.org/courses?level=Post+16" },
-    ];
-
-    teacherButtonDetails.forEach(({ text, href }, index) => {
-      expect(teacherDashboardButtons![index]).toHaveTextContent(text);
-      expect(teacherDashboardButtons![index]).toHaveAttribute("href", href);
-      const buttonIcon = within(teacherDashboardButtons![index]).getByRole("img");
-      expect(buttonIcon).toBeVisible();
-    });
-  };
-
-  expectCommonButtons();
-
-  if (role === "TEACHER") {
-    expectTeacherButtons();
-  } else {
-    expect(teacherDashboardButtons).toBeNull();
   }
 };
